♻️Refactor: 로직과 UI 부분 분리 작업 (상수화 파일 재정의, menuItem 컴포넌트 생성으로 인한 SidebarContent 간소화)
