+++ src/Components/common/menubar/MenuBar.tsx
@@ -1,27 +1,16 @@
-import { AdminMenuDesktop } from "./AdminMenuDesktop";
-import { AdminMenuMobile } from "./AdminMenuMobile";
+import { ManagerMenuBar } from "./ManagerMenuBar";
 import { EmployeeMenuBar } from "./EmployeeMenuBar";
 import { useMenuBar } from "@/hooks/menu/useMenuBar";
 
 const MenuBar = () => {
   const menuBar = useMenuBar();
-  const { userType, isDesktop, ...menuState } = menuBar;
+  const { userType, ...menuState } = menuBar;
 
   if (userType === "manager") {
-    if (isDesktop) {
-      return <AdminMenuDesktop {...menuState} />;
-    }
-    return (
-      <AdminMenuMobile
-        companyLogo={menuState.companyLogo}
-        companyName={menuState.companyName}
-        darkMode={menuState.darkMode}
-        logout={menuState.logout}
-      />
-    );
+    return <ManagerMenuBar {...menuState} />;
+  } else {
+    return <EmployeeMenuBar {...menuState} />;
   }
-
-  return <EmployeeMenuBar {...menuState} />;
 };
 
 export default MenuBar;
