+++ src/hooks/menu/useMenuBar.ts
@@ -1,12 +1,10 @@
 import { useLocation } from "react-router-dom";
-import { useWindowWidth } from "./useWindowWidth";
 import { useMenuState } from "./useMenuState";
 import { useMenuActions } from "./useMenuActions";
 import { useMenuItems } from "./useMenuItems";
 
 export const useMenuBar = () => {
   const location = useLocation();
-  const { isDesktop } = useWindowWidth();
   const menuState = useMenuState();
   const menuActions = useMenuActions();
   const menuItems = useMenuItems(menuState.companyCode, menuActions.logout);
@@ -16,6 +14,5 @@ export const useMenuBar = () => {
     ...menuActions,
     ...menuItems,
     location,
-    isDesktop,
   };
 };
