--- src/hooks/menu/useWindowWidth.ts
@@ -1,18 +1,0 @@
-import { useEffect, useState } from "react";
-
-export const useWindowWidth = () => {
-  const [windowWidth, setWindowWidth] = useState(window.innerWidth);
-
-  useEffect(() => {
-    const handleResize = () => {
-      setWindowWidth(window.innerWidth);
-    };
-
-    window.addEventListener("resize", handleResize);
-    return () => window.removeEventListener("resize", handleResize);
-  }, []);
-
-  return {
-    isDesktop: windowWidth > 600,
-  };
-};
